started the big rewrite
