+++ consts.py
@@ -7,3 +7,9 @@ INSTRUCTOR = "instructor"
 OK_ROLES = [STUDENT, LAB_ASSISTANT, GRADER, STAFF, INSTRUCTOR]
 
 NO_CAPTION_TRACK_MESSAGE = 'Please make sure that (a) you own this video, and (b) there is a caption track associated with this video: see https://support.google.com/youtube/answer/2734796?hl=en, and try again.'
+
+CLIENT_SECRETS_FILE = 'keys/client_secret.json'
+
+SCOPES = ['https://www.googleapis.com/auth/youtube.force-ssl']
+API_SERVICE_NAME = 'youtube'
+API_VERSION = 'v3'
